Fixes tester, merges with feature
Big changes here. Have cropped back previous progress to make room for
growth from other people with the idea that the more scripts are here,
the more confusing it is to get started. Will continue to work on the
documents for the research article, proposal for funding, and
recruitment attempts